+++ sdk/storage/azure-storage-files-shares/test/share_client_test.cpp
@@ -227,115 +227,112 @@ namespace Azure { namespace Storage { namespace Test {
     EXPECT_EQ(expectedPermission, ret2->Permission);
   }
 
-  // TEST_F(FileShareClientTest, Lease)
-  //{
-  //  std::string leaseId1 = CreateUniqueLeaseId();
-  //  int32_t leaseDuration = 20;
-  //  auto aLease = *m_shareClient->AcquireLease(leaseId1, leaseDuration);
-  //  EXPECT_FALSE(aLease.ETag.empty());
-  //  EXPECT_FALSE(aLease.LastModified.empty());
-  //  EXPECT_EQ(aLease.LeaseId, leaseId1);
-  //  aLease = *m_shareClient->AcquireLease(leaseId1, leaseDuration);
-  //  EXPECT_FALSE(aLease.ETag.empty());
-  //  EXPECT_FALSE(aLease.LastModified.empty());
-  //  EXPECT_EQ(aLease.LeaseId, leaseId1);
+  TEST_F(FileShareClientTest, Lease)
+  {
+    std::string leaseId1 = CreateUniqueLeaseId();
+    std::chrono::seconds leaseDuration(20);
+    auto leaseClient = Files::Shares::ShareLeaseClient(*m_shareClient, leaseId1);
 
-  //  auto properties = *m_shareClient->GetProperties();
-  //  EXPECT_EQ(properties.LeaseState.GetValue(), Files::Shares::LeaseStateType::Leased);
-  //  EXPECT_EQ(properties.LeaseStatus.GetValue(), Files::Shares::LeaseStatusType::Locked);
-  //  EXPECT_EQ(Files::Shares::LeaseDurationType::Fixed, properties.LeaseDuration.GetValue());
+    auto aLease = *leaseClient.Acquire(leaseDuration);
+    EXPECT_FALSE(aLease.ETag.empty());
+    EXPECT_NE(Azure::Core::DateTime(), aLease.LastModified);
+    EXPECT_EQ(aLease.LeaseId, leaseId1);
+    aLease = *leaseClient.Acquire(leaseDuration);
+    EXPECT_FALSE(aLease.ETag.empty());
+    EXPECT_NE(Azure::Core::DateTime(), aLease.LastModified);
+    EXPECT_EQ(aLease.LeaseId, leaseId1);
 
-  //  auto rLease = *m_shareClient->RenewLease(leaseId1);
-  //  EXPECT_FALSE(rLease.ETag.empty());
-  //  EXPECT_FALSE(rLease.LastModified.empty());
-  //  EXPECT_EQ(rLease.LeaseId, leaseId1);
+    auto properties = *m_shareClient->GetProperties();
+    EXPECT_EQ(properties.LeaseState.GetValue(), Files::Shares::Models::LeaseStateType::Leased);
+    EXPECT_EQ(properties.LeaseStatus.GetValue(), Files::Shares::Models::LeaseStatusType::Locked);
+    EXPECT_EQ(Files::Shares::Models::LeaseDurationType::Fixed, properties.LeaseDuration.GetValue());
 
-  //  std::string leaseId2 = CreateUniqueLeaseId();
-  //  EXPECT_NE(leaseId1, leaseId2);
-  //  auto cLease = *m_shareClient->ChangeLease(leaseId1, leaseId2);
-  //  EXPECT_FALSE(cLease.ETag.empty());
-  //  EXPECT_FALSE(cLease.LastModified.empty());
-  //  EXPECT_EQ(cLease.LeaseId, leaseId2);
+    auto rLease = *leaseClient.Renew();
+    EXPECT_FALSE(rLease.ETag.empty());
+    EXPECT_NE(Azure::Core::DateTime(), rLease.LastModified);
+    EXPECT_EQ(rLease.LeaseId, leaseId1);
 
-  //  auto blobInfo = *m_shareClient->ReleaseLease(leaseId2);
-  //  EXPECT_FALSE(blobInfo.ETag.empty());
-  //  EXPECT_FALSE(blobInfo.LastModified.empty());
+    std::string leaseId2 = CreateUniqueLeaseId();
+    EXPECT_NE(leaseId1, leaseId2);
+    auto cLease = *leaseClient.Change(leaseId2);
+    EXPECT_FALSE(cLease.ETag.empty());
+    EXPECT_NE(Azure::Core::DateTime(), cLease.LastModified);
+    EXPECT_EQ(cLease.LeaseId, leaseId2);
 
-  //  aLease = *m_shareClient->AcquireLease(CreateUniqueLeaseId(), InfiniteLeaseDuration);
-  //  properties = *m_shareClient->GetProperties();
-  //  EXPECT_EQ(Files::Shares::LeaseDurationType::Infinite, properties.LeaseDuration.GetValue());
-  //  auto brokenLease = *m_shareClient->BreakLease();
-  //  EXPECT_FALSE(brokenLease.ETag.empty());
-  //  EXPECT_FALSE(brokenLease.LastModified.empty());
-  //  EXPECT_EQ(brokenLease.LeaseTime, 0);
+    auto relLease = *leaseClient.Release();
+    EXPECT_FALSE(relLease.ETag.empty());
+    EXPECT_NE(Azure::Core::DateTime(), relLease.LastModified);
 
-  //  aLease = *m_shareClient->AcquireLease(CreateUniqueLeaseId(), leaseDuration);
-  //  brokenLease = *m_shareClient->BreakLease();
-  //  EXPECT_FALSE(brokenLease.ETag.empty());
-  //  EXPECT_FALSE(brokenLease.LastModified.empty());
-  //  EXPECT_NE(brokenLease.LeaseTime, 0);
+    leaseClient = Files::Shares::ShareLeaseClient(*m_shareClient, CreateUniqueLeaseId());
+    aLease = *leaseClient.Acquire(Files::Shares::ShareLeaseClient::InfiniteLeaseDuration);
+    properties = *m_shareClient->GetProperties();
+    EXPECT_EQ(
+        Files::Shares::Models::LeaseDurationType::Infinite, properties.LeaseDuration.GetValue());
+    auto brokenLease = *leaseClient.Break();
+    EXPECT_FALSE(brokenLease.ETag.empty());
+    EXPECT_NE(Azure::Core::DateTime(), brokenLease.LastModified);
+    EXPECT_EQ(brokenLease.LeaseTime, 0);
 
-  //  Files::Shares::BreakShareLeaseOptions options;
-  //  options.BreakPeriod = 0;
-  //  m_shareClient->BreakLease(options);
-  //}
+    Files::Shares::BreakShareLeaseOptions options;
+    options.BreakPeriod = 0;
+    leaseClient.Break(options);
+  }
 
-  // TEST_F(FileShareClientTest, SnapshotLease)
-  //{
-  //  std::string leaseId1 = CreateUniqueLeaseId();
-  //  int32_t leaseDuration = 20;
-  //  auto snapshotResult = m_shareClient->CreateSnapshot();
-  //  auto shareSnapshot = m_shareClient->WithSnapshot(snapshotResult->Snapshot);
-  //  auto aLease = *shareSnapshot.AcquireLease(leaseId1, leaseDuration);
-  //  EXPECT_FALSE(aLease.ETag.empty());
-  //  EXPECT_FALSE(aLease.LastModified.empty());
-  //  EXPECT_EQ(aLease.LeaseId, leaseId1);
-  //  aLease = *shareSnapshot.AcquireLease(leaseId1, leaseDuration);
-  //  EXPECT_FALSE(aLease.ETag.empty());
-  //  EXPECT_FALSE(aLease.LastModified.empty());
-  //  EXPECT_EQ(aLease.LeaseId, leaseId1);
+  TEST_F(FileShareClientTest, SnapshotLease)
+  {
+    std::string leaseId1 = CreateUniqueLeaseId();
+    std::chrono::seconds leaseDuration(20);
+    auto snapshotResult = m_shareClient->CreateSnapshot();
+    auto shareSnapshot = m_shareClient->WithSnapshot(snapshotResult->Snapshot);
+    auto shareSnapshotLeaseClient = Files::Shares::ShareLeaseClient(shareSnapshot, leaseId1);
+    auto aLease = *shareSnapshotLeaseClient.Acquire(leaseDuration);
+    EXPECT_FALSE(aLease.ETag.empty());
+    EXPECT_NE(Azure::Core::DateTime(), aLease.LastModified);
+    EXPECT_EQ(aLease.LeaseId, leaseId1);
+    aLease = *shareSnapshotLeaseClient.Acquire(leaseDuration);
+    EXPECT_FALSE(aLease.ETag.empty());
+    EXPECT_NE(Azure::Core::DateTime(), aLease.LastModified);
+    EXPECT_EQ(aLease.LeaseId, leaseId1);
 
-  //  auto properties = *shareSnapshot.GetProperties();
-  //  EXPECT_EQ(properties.LeaseState.GetValue(), Files::Shares::LeaseStateType::Leased);
-  //  EXPECT_EQ(properties.LeaseStatus.GetValue(), Files::Shares::LeaseStatusType::Locked);
-  //  EXPECT_EQ(Files::Shares::LeaseDurationType::Fixed, properties.LeaseDuration.GetValue());
+    auto properties = *shareSnapshot.GetProperties();
+    EXPECT_EQ(properties.LeaseState.GetValue(), Files::Shares::Models::LeaseStateType::Leased);
+    EXPECT_EQ(properties.LeaseStatus.GetValue(), Files::Shares::Models::LeaseStatusType::Locked);
+    EXPECT_EQ(Files::Shares::Models::LeaseDurationType::Fixed, properties.LeaseDuration.GetValue());
 
-  //  auto rLease = *shareSnapshot.RenewLease(leaseId1);
-  //  EXPECT_FALSE(rLease.ETag.empty());
-  //  EXPECT_FALSE(rLease.LastModified.empty());
-  //  EXPECT_EQ(rLease.LeaseId, leaseId1);
+    auto rLease = *shareSnapshotLeaseClient.Renew();
+    EXPECT_FALSE(rLease.ETag.empty());
+    EXPECT_NE(Azure::Core::DateTime(), rLease.LastModified);
+    EXPECT_EQ(rLease.LeaseId, leaseId1);
 
-  //  std::string leaseId2 = CreateUniqueLeaseId();
-  //  EXPECT_NE(leaseId1, leaseId2);
-  //  auto cLease = *shareSnapshot.ChangeLease(leaseId1, leaseId2);
-  //  EXPECT_FALSE(cLease.ETag.empty());
-  //  EXPECT_FALSE(cLease.LastModified.empty());
-  //  EXPECT_EQ(cLease.LeaseId, leaseId2);
+    std::string leaseId2 = CreateUniqueLeaseId();
+    EXPECT_NE(leaseId1, leaseId2);
+    auto cLease = *shareSnapshotLeaseClient.Change(leaseId2);
+    EXPECT_FALSE(cLease.ETag.empty());
+    EXPECT_NE(Azure::Core::DateTime(), cLease.LastModified);
+    EXPECT_EQ(cLease.LeaseId, leaseId2);
 
-  //  auto blobInfo = *shareSnapshot.ReleaseLease(leaseId2);
-  //  EXPECT_FALSE(blobInfo.ETag.empty());
-  //  EXPECT_FALSE(blobInfo.LastModified.empty());
+    auto relLease = *shareSnapshotLeaseClient.Release();
+    EXPECT_FALSE(relLease.ETag.empty());
+    EXPECT_NE(Azure::Core::DateTime(), relLease.LastModified);
 
-  //  aLease = *shareSnapshot.AcquireLease(CreateUniqueLeaseId(), InfiniteLeaseDuration);
-  //  properties = *shareSnapshot.GetProperties();
-  //  EXPECT_EQ(Files::Shares::LeaseDurationType::Infinite, properties.LeaseDuration.GetValue());
-  //  auto brokenLease = *shareSnapshot.BreakLease();
-  //  EXPECT_FALSE(brokenLease.ETag.empty());
-  //  EXPECT_FALSE(brokenLease.LastModified.empty());
-  //  EXPECT_EQ(brokenLease.LeaseTime, 0);
+    shareSnapshotLeaseClient
+        = Files::Shares::ShareLeaseClient(shareSnapshot, CreateUniqueLeaseId());
+    aLease
+        = *shareSnapshotLeaseClient.Acquire(Files::Shares::ShareLeaseClient::InfiniteLeaseDuration);
+    properties = *shareSnapshot.GetProperties();
+    EXPECT_EQ(
+        Files::Shares::Models::LeaseDurationType::Infinite, properties.LeaseDuration.GetValue());
+    auto brokenLease = *shareSnapshotLeaseClient.Break();
+    EXPECT_FALSE(brokenLease.ETag.empty());
+    EXPECT_NE(Azure::Core::DateTime(), brokenLease.LastModified);
+    EXPECT_EQ(brokenLease.LeaseTime, 0);
 
-  //  aLease = *shareSnapshot.AcquireLease(CreateUniqueLeaseId(), leaseDuration);
-  //  brokenLease = *shareSnapshot.BreakLease();
-  //  EXPECT_FALSE(brokenLease.ETag.empty());
-  //  EXPECT_FALSE(brokenLease.LastModified.empty());
-  //  EXPECT_NE(brokenLease.LeaseTime, 0);
+    Files::Shares::BreakShareLeaseOptions options;
+    options.BreakPeriod = 0;
+    shareSnapshotLeaseClient.Break(options);
 
-  //  Files::Shares::BreakShareLeaseOptions options;
-  //  options.BreakPeriod = 0;
-  //  shareSnapshot.BreakLease(options);
-
-  //  EXPECT_THROW(m_shareClient->Delete(), StorageException);
-  //}
+    EXPECT_THROW(m_shareClient->Delete(), StorageException);
+  }
 
   TEST_F(FileShareClientTest, UnencodedDirectoryFileNameWorks)
   {