+++ sdk/storage/azure-storage-files-shares/CHANGELOG.md
@@ -11,6 +11,7 @@
 - Moved all protocol layer generated result types to `Details` namespace.
 - Renamed `ShareItems` in `ListSharesResponse` to `Items`.
 - Renamed `ShareItems` in `ServiceListSharesSinglePageResult` to `Items`.
+- Added `ShareLeaseClient`, all lease related APIs are moved to `ShareLeaseClient`.
 
 ## 12.0.0-beta.6 (2020-01-14)